+++ components/CodeBlock/code-block-copy-button.tsx
@@ -1,6 +1,7 @@
 'use client'
 
 import { useCallback, useState } from 'react'
+import copy from 'copy-to-clipboard'
 import { CheckIcon, CopyIcon } from './icons'
 
 const COPY_LABEL_WIDTH_PX = 68
@@ -17,14 +18,11 @@ export function CodeBlockCopyButton({
 }) {
   const [copied, setCopied] = useState(false)
 
-  const onCopy = useCallback(async () => {
-    try {
-      await navigator.clipboard.writeText(text)
-      setCopied(true)
-      window.setTimeout(() => setCopied(false), 2000)
-    } catch {
-      // Clipboard can fail in insecure contexts; fail silently.
-    }
+  const onCopy = useCallback(() => {
+    const ok = copy(text)
+    if (!ok) return
+    setCopied(true)
+    window.setTimeout(() => setCopied(false), 2000)
   }, [text])
 
   return (
