+++ components/CodeBlock/code-block.module.css
@@ -312,6 +312,26 @@
   line-height: inherit !important;
   white-space: pre;
   tab-size: 2;
+  scrollbar-width: thin;
+  scrollbar-color: var(--code-block-border, rgba(255, 255, 255, 0.2)) transparent;
+}
+
+.pre::-webkit-scrollbar {
+  width: 8px;
+  height: 8px;
+}
+
+.pre::-webkit-scrollbar-track {
+  background: transparent;
+}
+
+.pre::-webkit-scrollbar-thumb {
+  border-radius: 4px;
+  background-color: var(--code-block-border, rgba(255, 255, 255, 0.25));
+}
+
+.pre::-webkit-scrollbar-thumb:hover {
+  background-color: var(--code-block-foreground, rgba(255, 255, 255, 0.4));
 }
 
 .pre :global(code) {
