+++ package.json
@@ -42,6 +42,7 @@
     "@vimeo/player": "^2.23.0",
     "algoliasearch": "^5.37.0",
     "autoprefixer": "^10.4.13",
+    "copy-to-clipboard": "^4.0.2",
     "date-fns": "^3.6.0",
     "dompurify": "^3.4.12",
     "framer-motion": "^11.13.5",
